// gantryweb/Frontend/Gantry/src/app/cds/matches/boxing-cds/models/boxing-cds-constants.model.ts
export enum BoxingCdsTemplateIds {
  FIGHTBETTING = "108",
  ROUNDBETTING = "11380",
  ROUNDGROUPBETTING = "31367",
  METHODOFVICTORY = "23458"
}
export enum RoundBetting {
  POINTSNAME = 'on points',
  ROUND = 'round',
  BETNAME = 'betName',
}

export enum SelectionName {
  POINTSNAME = '(FULL DISTANCE)'
}

export const titleFilter = {
  TechnicalDecision: "KO/TKO/Technical Decision or DQ",
  OnPoints: "On points"
}